pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatch-vcs",
  "hatchling>=1.27",
]

[project]
name = "pylast"
description = "A Python interface to Last.fm and Libre.fm"
readme = "README.md"
keywords = [
  "Last.fm",
  "music",
  "scrobble",
  "scrobbling",
]
license = "Apache-2.0"
license-files = [ "LICENSE.txt" ]
maintainers = [ { name = "Hugo van Kemenade" } ]
authors = [ { name = "Amr Hassan <amr.hassan@gmail.com> and Contributors", email = "amr.hassan@gmail.com" } ]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: 3.15",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
  "Topic :: Internet",
  "Topic :: Multimedia :: Sound/Audio",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
dynamic = [ "version" ]
dependencies = [
  "httpx>=0.26",
]
optional-dependencies.tests = [
  "flaky",
  "pytest>=9",
  "pytest-cov",
  "pytest-random-order",
  "pytest-recording",
  "pyyaml",
]
urls.Changelog = "https://github.com/pylast/pylast/releases"
urls.Homepage = "https://github.com/pylast/pylast"
urls.Source = "https://github.com/pylast/pylast"

[tool.hatch]
version.source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/pylast/_version.py"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.ruff]
fix = true

lint.select = [
  "C4",     # flake8-comprehensions
  "E",      # pycodestyle errors
  "EM",     # flake8-errmsg
  "F",      # pyflakes
  "I",      # isort
  "ICN",    # flake8-import-conventions
  "ISC",    # flake8-implicit-str-concat
  "LOG",    # flake8-logging
  "PGH",    # pygrep-hooks
  "PIE",    # flake8-pie
  "PT",     # flake8-pytest-style
  "PYI",    # flake8-pyi
  "RUF022", # unsorted-dunder-all
  "RUF100", # unused noqa (yesqa)
  "UP",     # pyupgrade
  "W",      # pycodestyle warnings
  "YTT",    # flake8-2020
]
lint.ignore = [
  "E203",   # Whitespace before ':'
  "E221",   # Multiple spaces before operator
  "E226",   # Missing whitespace around arithmetic operator
  "E241",   # Multiple spaces after ','
  "PIE790", # Unnecessary pass statement
]
lint.flake8-import-conventions.aliases.datetime = "dt"
lint.flake8-import-conventions.banned-from = [ "datetime" ]
lint.flake8-pytest-style.parametrize-names-type = "csv"
lint.isort.known-first-party = [ "pylast" ]
lint.isort.required-imports = [ "from __future__ import annotations" ]

[tool.pyproject-fmt]
max_supported_python = "3.15"

[tool.pytest]
minversion = "9.0"
filterwarnings = [
  "once::DeprecationWarning",
  "once::PendingDeprecationWarning",
]
testpaths = [ "tests" ]
xfail_strict = true

[tool.mypy]
pretty = true
show_error_codes = true
